refactor: Swap accessors from inheritance to composition
