+++ bigframes/core/indexes/base.py
@@ -384,13 +384,13 @@ class Index(vendored_pandas_index.Index):
         name = self.name if name is None else name
         if index is None:
             return bigframes.series.Series(
-                data=self, index=self, name=name, session=self._session
+                data=self, index=self, name=str(name), session=self._session
             )
         else:
             return bigframes.series.Series(
                 data=self,
                 index=Index(index, session=self._session),
-                name=name,
+                name=str(name),
                 session=self._session,
             )
