+++ bigframes/operations/strings.py
@@ -25,7 +25,6 @@ import bigframes.dataframe as df
 import bigframes.operations as ops
 from bigframes.operations._op_converters import convert_index, convert_slice
 import bigframes.operations.aggregations as agg_ops
-import bigframes.operations.base
 import bigframes.series as series
 
 # Maps from python to re2
@@ -37,14 +36,17 @@ REGEXP_FLAGS = {
 
 
 @log_adapter.class_logger
-class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMethods):
+class StringMethods(vendorstr.StringMethods):
     __doc__ = vendorstr.StringMethods.__doc__
+
+    def __init__(self, data: series.Series):
+        self._data = data
 
     def __getitem__(self, key: Union[int, slice]) -> series.Series:
         if isinstance(key, int):
-            return self._apply_unary_op(convert_index(key))
+            return self._data._apply_unary_op(convert_index(key))
         elif isinstance(key, slice):
-            return self._apply_unary_op(convert_slice(key))
+            return self._data._apply_unary_op(convert_slice(key))
         else:
             raise ValueError(f"key must be an int or slice, got {type(key).__name__}")
 
@@ -54,13 +56,15 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
         start: Optional[int] = None,
         end: Optional[int] = None,
     ) -> series.Series:
-        return self._apply_unary_op(ops.StrFindOp(substr=sub, start=start, end=end))
+        return self._data._apply_unary_op(
+            ops.StrFindOp(substr=sub, start=start, end=end)
+        )
 
     def len(self) -> series.Series:
-        return self._apply_unary_op(ops.len_op)
+        return self._data._apply_unary_op(ops.len_op)
 
     def lower(self) -> series.Series:
-        return self._apply_unary_op(ops.lower_op)
+        return self._data._apply_unary_op(ops.lower_op)
 
     def reverse(self) -> series.Series:
         """Reverse strings in the Series.
@@ -83,76 +87,76 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
                 pattern matches the start of each string element.
         """
         # reverse method is in ibis, not pandas.
-        return self._apply_unary_op(ops.reverse_op)
+        return self._data._apply_unary_op(ops.reverse_op)
 
     def slice(
         self,
         start: Optional[int] = None,
         stop: Optional[int] = None,
     ) -> series.Series:
-        return self._apply_unary_op(ops.StrSliceOp(start=start, end=stop))
+        return self._data._apply_unary_op(ops.StrSliceOp(start=start, end=stop))
 
     def strip(self, to_strip: Optional[str] = None) -> series.Series:
-        return self._apply_unary_op(
+        return self._data._apply_unary_op(
             ops.StrStripOp(to_strip=" \n\t" if to_strip is None else to_strip)
         )
 
     def upper(self) -> series.Series:
-        return self._apply_unary_op(ops.upper_op)
+        return self._data._apply_unary_op(ops.upper_op)
 
     def isnumeric(self) -> series.Series:
-        return self._apply_unary_op(ops.isnumeric_op)
+        return self._data._apply_unary_op(ops.isnumeric_op)
 
     def isalpha(
         self,
     ) -> series.Series:
-        return self._apply_unary_op(ops.isalpha_op)
+        return self._data._apply_unary_op(ops.isalpha_op)
 
     def isdigit(
         self,
     ) -> series.Series:
-        return self._apply_unary_op(ops.isdigit_op)
+        return self._data._apply_unary_op(ops.isdigit_op)
 
     def isdecimal(
         self,
     ) -> series.Series:
-        return self._apply_unary_op(ops.isdecimal_op)
+        return self._data._apply_unary_op(ops.isdecimal_op)
 
     def isalnum(
         self,
     ) -> series.Series:
-        return self._apply_unary_op(ops.isalnum_op)
+        return self._data._apply_unary_op(ops.isalnum_op)
 
     def isspace(
         self,
     ) -> series.Series:
-        return self._apply_unary_op(ops.isspace_op)
+        return self._data._apply_unary_op(ops.isspace_op)
 
     def islower(
         self,
     ) -> series.Series:
-        return self._apply_unary_op(ops.islower_op)
+        return self._data._apply_unary_op(ops.islower_op)
 
     def isupper(
         self,
     ) -> series.Series:
-        return self._apply_unary_op(ops.isupper_op)
+        return self._data._apply_unary_op(ops.isupper_op)
 
     def rstrip(self, to_strip: Optional[str] = None) -> series.Series:
-        return self._apply_unary_op(
+        return self._data._apply_unary_op(
             ops.StrRstripOp(to_strip=" \n\t" if to_strip is None else to_strip)
         )
 
     def lstrip(self, to_strip: Optional[str] = None) -> series.Series:
-        return self._apply_unary_op(
+        return self._data._apply_unary_op(
             ops.StrLstripOp(to_strip=" \n\t" if to_strip is None else to_strip)
         )
 
     def repeat(self, repeats: int) -> series.Series:
-        return self._apply_unary_op(ops.StrRepeatOp(repeats=repeats))
+        return self._data._apply_unary_op(ops.StrRepeatOp(repeats=repeats))
 
     def capitalize(self) -> series.Series:
-        return self._apply_unary_op(ops.capitalize_op)
+        return self._data._apply_unary_op(ops.capitalize_op)
 
     def match(self, pat, case=True, flags=0) -> series.Series:
         # \A anchors start of entire string rather than start of any line in multiline mode
@@ -166,20 +170,20 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
         return self.contains(pat=adj_pat, case=case, flags=flags)
 
     def get(self, i: int) -> series.Series:
-        return self._apply_unary_op(ops.StrGetOp(i=i))
+        return self._data._apply_unary_op(ops.StrGetOp(i=i))
 
     def pad(self, width, side="left", fillchar=" ") -> series.Series:
-        return self._apply_unary_op(
+        return self._data._apply_unary_op(
             ops.StrPadOp(length=width, fillchar=fillchar, side=side)
         )
 
     def ljust(self, width, fillchar=" ") -> series.Series:
-        return self._apply_unary_op(
+        return self._data._apply_unary_op(
             ops.StrPadOp(length=width, fillchar=fillchar, side="right")
         )
 
     def rjust(self, width, fillchar=" ") -> series.Series:
-        return self._apply_unary_op(
+        return self._data._apply_unary_op(
             ops.StrPadOp(length=width, fillchar=fillchar, side="left")
         )
 
@@ -192,9 +196,9 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
             re2flags = _parse_flags(flags)
             if re2flags:
                 pat = re2flags + pat
-            return self._apply_unary_op(ops.StrContainsRegexOp(pat=pat))
+            return self._data._apply_unary_op(ops.StrContainsRegexOp(pat=pat))
         else:
-            return self._apply_unary_op(ops.StrContainsOp(pat=pat))
+            return self._data._apply_unary_op(ops.StrContainsOp(pat=pat))
 
     def extract(self, pat: str, flags: int = 0) -> df.DataFrame:
         re2flags = _parse_flags(flags)
@@ -205,7 +209,7 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
             raise ValueError("No capture groups in 'pat'")
 
         results: list[str] = []
-        block = self._block
+        block = self._data._block
         for i in range(compiled.groups):
             labels = [
                 label
@@ -214,7 +218,7 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
             ]
             label = labels[0] if labels else str(i)
             block, id = block.apply_unary_op(
-                self._value_column,
+                self._data._value_column,
                 ops.StrExtractOp(pat=pat, n=i + 1),
                 result_label=label,
             )
@@ -244,13 +248,15 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
             re2flags = _parse_flags(flags)
             if re2flags:
                 pat_str = re2flags + pat_str
-            return self._apply_unary_op(ops.RegexReplaceStrOp(pat=pat_str, repl=repl))
+            return self._data._apply_unary_op(
+                ops.RegexReplaceStrOp(pat=pat_str, repl=repl)
+            )
         else:
             if isinstance(pat, re.Pattern):
                 raise ValueError(
                     "Must set 'regex'=True if using compiled regex pattern."
                 )
-            return self._apply_unary_op(ops.ReplaceStrOp(pat=pat_str, repl=repl))
+            return self._data._apply_unary_op(ops.ReplaceStrOp(pat=pat_str, repl=repl))
 
     def startswith(
         self,
@@ -258,7 +264,7 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
     ) -> series.Series:
         if not isinstance(pat, tuple):
             pat = (pat,)
-        return self._apply_unary_op(ops.StartsWithOp(pat=pat))
+        return self._data._apply_unary_op(ops.StartsWithOp(pat=pat))
 
     def endswith(
         self,
@@ -266,7 +272,7 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
     ) -> series.Series:
         if not isinstance(pat, tuple):
             pat = (pat,)
-        return self._apply_unary_op(ops.EndsWithOp(pat=pat))
+        return self._data._apply_unary_op(ops.EndsWithOp(pat=pat))
 
     def split(
         self,
@@ -278,13 +284,13 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
                 "Regular expressions aren't currently supported. Please set "
                 + f"`regex=False` and try again. {constants.FEEDBACK_LINK}"
             )
-        return self._apply_unary_op(ops.StringSplitOp(pat=pat))
+        return self._data._apply_unary_op(ops.StringSplitOp(pat=pat))
 
     def zfill(self, width: int) -> series.Series:
-        return self._apply_unary_op(ops.ZfillOp(width=width))
+        return self._data._apply_unary_op(ops.ZfillOp(width=width))
 
     def center(self, width: int, fillchar: str = " ") -> series.Series:
-        return self._apply_unary_op(
+        return self._data._apply_unary_op(
             ops.StrPadOp(length=width, fillchar=fillchar, side="both")
         )
 
@@ -294,10 +300,10 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
         *,
         join: Literal["outer", "left"] = "left",
     ) -> series.Series:
-        return self._apply_binary_op(others, ops.strconcat_op, alignment=join)
+        return self._data._apply_binary_op(others, ops.strconcat_op, alignment=join)
 
     def join(self, sep: str) -> series.Series:
-        return self._apply_unary_op(
+        return self._data._apply_unary_op(
             ops.ArrayReduceOp(aggregation=agg_ops.StringAggOp(sep=sep))
         )
 
@@ -321,9 +327,9 @@ class StringMethods(bigframes.operations.base.SeriesMethods, vendorstr.StringMet
             bigframes.series.Series: Blob Series.
 
         """
-        session = self._block.session
+        session = self._data._block.session
         connection = session._create_bq_connection(connection=connection)
-        return self._apply_binary_op(connection, ops.obj_make_ref_op)
+        return self._data._apply_binary_op(connection, ops.obj_make_ref_op)
 
 
 def _parse_flags(flags: int) -> Optional[str]:
